Introduce capi machine API and nodePools
This drops the openshift machine API for guestcluster compute nodes in favour of using the upstream capi machine API.
It also introduces a new CRD nodePool which underneath leverages the capi machine scalable resources.
As elaborated in the readme this uses [1]/[2] on nodes given the current lack of a cloud provider. It also uses [3]/[4] to let the capi machine controller work with an external infrastructure resource.

[1] https://github.com/kubernetes-sigs/cluster-api-provider-aws/compare/master...enxebre:decouple-machine-infra to set providerIDs
[2] https://github.com/enxebre/cluster-api-provider-aws-2/tree/decouple-machine-infra]
[3] https://github.com/kubernetes-sigs/cluster-api/compare/master...enxebre:set-providerID?expand=1
[4] https://github.com/enxebre/cluster-api/tree/set-providerID