Improve installation
This commit introduces some major refactoring to make installation easier.

* Get rid of the `manifests` directory and handle all configuration through
Kustomize.

* Add a deployment (and all supporting resources) for the operator itself.

* Introduce reasonable defaults for discovering the control plane image to use
based on the operator pod (with the ability to override with a flag).

* Abstract away the release image metadata handling and integrate it with
Kustomize. For now all guest clusters will use same release version as the
management cluster, but the commit also also cleans up some of the plugin points
to help us improve in a followup.

* Clean up and improve the Makefile to be the entrypoint for all of these tasks.

* Simplify the installation procedures in the README.