Fix openshiftCluster deletion
This lets the nodePool controller to explicitly delete its owned machineSet. Removing the owner ref is not enough as it has other from the cluster resource.

It also move the default nodePool to be created by the openshiftCluster controller. Otherwise openshiftCluster deletion is not able to delete de default nodePool.