Rename control-plane-operator to hosted-cluster-config-operator
The control-plane-operator is responsible only for managing certain parts
of the configuration of a hosted control plane rather than the control plane
itself as a whole. Rename the operator to reflect its responsibilities and
allow the control-plane-operator name to be reclaimed by a new component
which will assume those broader responsibilities.