Introduce new managed Control Plane Operator component
Split out a new Control Plane Operator which is managed by the HyperShift
Operator.

The Control Plane Operator is a component deployed into each hosted cluster's
namespace that manages the rollout and upgrade of a single version of the the
cluster's control plane. Its key responsibilities are:

- Provisioning all the infrastructure required to host a control plane (whether
this means creating or adopting existing infrastructure). This infrastructure
may be management cluster resources, external cloud provider resources, etc.

- Deploying an OCP control plane (into the operator's namespace), configured to
run in the context of the provisioned infrastructure.

- Owning and implementing any behavior necessary to rollout (upgrade) the new
version (e.g. version specific changes at layers above OCP itself, e.g.
configuration or infrastructure changes).

The Control Plane Operator is versioned in line with OCP versions it knows how
to deploy and manage and is not coupled to the management cluster's version.