Replace Kustomize with a new hypershift CLI
Kustomize was being used for:

1. Installation
2. Example cluster generation

In these narrow use cases, it's not clear Kustomize makes things easier, and it
also propagates dependencies downstream as we move into the CI system. This
commit replaces Kustomize entirely by providing a new `hypershift` CLI tool which
knows how to install the application and also generate example clusters.

This approach should enable a more portable integration point with CI and also
remove dependencies for users wanting to kick the tires of HyperShift. Here are
some other ways users can take advantage:

1. Use the `hypershift` binary directly through a release or (potentially) `go get`
2. Run `hypershift` using a container engine
3. Run `hypershift` in an OpenShift cluster using `oc run`

Another side effect of the move to a 100% native Go solution is that the fixtures
representing things like an example cluster can be more clearly shared with the e2e
test infrastructure, which this commit also implements.

As a further simplification and cleanup, the external Ginkgo binary is removed
and the `go test` tool is now the entrypoint for all testing.

Lastly, this commit adds support for looking up the OCP release image
so that reliable defaults can be used in all contexts (while providing easy
flags to override it.)