build: optimize code generation
This patch introduces a new `api` make target which should be run whenever code
in the `openshift.io/hypershift/api` package changes. It regenerates things like
deepcopy boilerplate and CRD manifest files. Overall this means binary builds
are faster as we don't run controller-gen unnecessarily.

There's also a hack to work around controller-gen's behavior that prepends the
CRD YAML with document separators which break deserialization. The hack makes it
so we can live with the raw output of controller-gen without post-processing it
without dealing with additional shell work.

In a followup we should introduce a step to the `verify` target which detects
missed `api` executions for PRs.