Add nodePoolManagement API
This extends the nodePool API to support rolling upgrades.
The future implementation will be:
An ad-hoc controller that watches the hostedCluster CR and it reconciles when it sees a new version i.e status.Version.
Reconciliation upgrades nodePools one by one:

Fetch all nodePools and filter out those with autoUpgrade=false.
Get first nodePool, disable autoscaling if enabled, modify the machineDeployment spec to point to the new user-data-${version} (and in futre ami) generated by the controlPlane upgrade.
Wait for machineDepolyment to finish the rolling upgrade. Move to the next machineDeployment.