Implement basic control plane upgrade orchestration
This commit introduces basic support for control plane upgrades according to the
HostedCluster versioning API.

Control plane upgrades are initiated by updating the HostedCluster
`.spec.release.image` field, which is propagated to the HostedControlPlane
resource following some rules which are articulated in the test cases but not
yet codified in the API. For example, rollouts are serialized. See the test
cases for more details.

To make this work, the control plane operator has been updated to remove gating
logic which caused it to stop reconciling in reaction to fields like a "ready"
flag (which are reminiscent of state machine behavior). Now the
HostedControlPlane will always be reconciled idempotently.

The HostedControlPlane controller itself isn't architected with granular
reconciliation in mind and so there are still some spurious updates and
reconcile attempts when a no-op would have been preferred, but things should be
basically correct.

In addition, the HostedControlPlane controller is not very accurate with regards
to its status reporting, as it does no deep per-component analysis to gauge
control plane health. This weakness of reporting is reflected to some extent now
in HostedCluster (garbage in, garbage out).

This commit expands status for HostedCluster with a basic Availability condition
(see the new tests for details).

Most of the status reporting happening now is very minimal and has lots of room
for improvement to provide better decision making context for other logic. The
goal of this commit is to establish the bare minimum parts to make everything
work together and demonstrate a functional system for iteration.