e2e: plumb a cancellable context through the e2e code
The test process should handle graceful shutdown in response to a
SIGTERM or SIGINT. To achieve this, provide each Test* function access
to a global test context.Context which will be closed upon receipt
of SIGINT or SIGTERM on the main test thread. This gives test functions
an opportunity to execute deferred calls.