cli: Add a `create kubeconfig` command to help working with control planes
This commit introduces a `create kubeconfig` command to the CLI which renders a
kubeconfig with a context for every HostedCluster resource. The kubeconfig for
each cluster is based on the secret referenced by the status of the
HostedCluster.

This is a helper meant to be incorporated into other CLI workflows.