Implement version history completion time
This commit implements another missing piece of the HostedCluster version API to
track the completion time of rollouts. The HostedControlPlane is responsible for
tracking the transition time of its image rollouts and now reflects that on its
status using a new LastReleaseImageTransitionTime field. That transition time is
then mirrored to the HostedCluster during version the status calculation when
updating history entries.