cli: destroy infra after cluster deletion grace period expires
This commit adds a cluster deletion grace period option to the `destroy cluster`
command so that if the hostedcluster fails to be deleted for whatever reason the
infra can still be torn down to prevent leaks. The default grace period is 10
minutes, but can be configured for whatever the use case demands.