cli: Enable using separate AWS credentials for clusters
This commit introduces a `--cluster-aws-creds` option to the `create cluster`
command which allows a separate AWS credentials file to be provided for use by
the HostedCluster distinct from the credentials used for infrastructure
creation. If `--cluster-aws-creds` is left blank, `--aws-creds` value is used by
default, which is the current behavior before this commit.