Separate cloud credentials by context
The KCM and CAPI components have separate responsibilities and different
permission needs. Each of these components should use distinct AWS IAM policies
that we define and support so that users have more control over the exposed
surface area of accounts attached to those policies.

This commit introduces the API to support this policy separation but still
uses a single credentials file for both contexts as a first backwards compatible
step. A followup commit can introduce support for users to provide separate
credentials and/or minting new ones as part of the CLI tooling.