Use CloudFormations for AWS infrastructure
This commit totally rewrites the AWS infrastructure handling to use AWS
CloudFormations, replacing all the bespoke AWS API code, because it has grown
beyond our ability to maintain, and AWS can do most of this work for us.
CloudFormations also presents a more clear and declarative infrastructure
contract for end users through the CloudFormations template construct.

The main objective refactor is to achieve functional parity with the existing
behavior, but implemented in CloudFormations. This commit is totally compatible
with the existing CLI/documentation.

There are some notable structural changes introduced:

- DNS infrastruture is now better isolated. Instead of using the base domain's
zone as the public zone for the hosted cluster, a new public zone is provisioned
for the cluster and an NS record is installed into the base domain zone which
delegates to the cluster's public zone. This means the hosted cluster only knows
about zones which are designated specifically for that cluster and can no longer
leak records into the base domain zone. It also means that when cleaning up, we
only need to track the NS record in the base domain zone and can safely delete
everything in the hosted cluster's zones.

- The code itself is restructured to no longer distinguish between IAM and other
infra, with the intent of separating IAM out in a followup after a deliberate
audit of our policy needs. This means the `create iam` command is no longer
necessary and has been deleted.

- The CLI is refactored to make a clearer distinction between provisioning (or
receiving) infrastructure and installing supporting resources into that
infrastructure (e.g. OIDC resources installed into an S3 bucket, or installing
an NS record into a zone).