Use bespoke, narrowly scoped AWS users for guest clusters
Before this commit, the admin's AWS credentials passed to `create cluster` were
copied into the guest cluster for use by each credentials context. The AWS
credentials passed to `create cluster` are assumed to be super-privileged and
not intended to be leaked into guest clusters.

Now, when creating a new cluster, OCP is given credentials for users created
specifically for that cluster as part of infra creation.

This is another step towards IAM isolation by introducing new policies for each
HostedCluster credentials context (KubeCloudControllerCreds and
NodePoolManagementCreds) and creating new users associated with each of those
policies. The CloudFormation template takes care of creating the policies and
users and exposes the credentials as outputs.