Bring OLM to control-plane namespace
Also removes OLM resources from the CVO manifests as part of the cvo
init process so they are not created on the guest cluster.

Notable future work:
1) uses dns/vpn configuration to run the default catalog pods in the control plane
while making their grpc api available to the guest cluster.  This
will likely move to konnectivity in the future

2) the set of default catalog pods cannot currently be controlled, in
the future this needs to be controllerable by the deployer of the
guest cluster

3) no olm metric endpoints+alerting rules are currently defined