cli: Add AWS Go SDK client retry handling tweaks
This commit cleans up the AWS client handling throughout the CLI. From a given
entrypoint, AWS clients are now created exactly once and shared amongst all
components.

* Enables all the `create` and `destroy` commands to accept preallocated client
instances from the caller.
* Lifts all client creation up to the nearest entrypoint of each usage context
and injects them downwards from there.
* Consolidates all session and config creation into a single utilty package.
* Sets some new defaults for the SDK's built-in retry logic.

Further work should be done to tweak the SDK's built-in retry facilities, and
that can now be done centrally to start with and individual services can be
further tweaked as issues are identified.