Add a hard timeout to release image metadata lookups
Impose a hard-coded 2 minute timeout for release image metadata lookups
so the reconciliation can't hang if lookups take too long. If this value
proves to be flaky we can adjust it later, but the current indefinite wait
causes destroy to hang forever if, for example, the lookup image can't
be pulled.