e2e: Improve cluster destroy reliability
Before this commit, if the e2e test failed to created the HostedCluster
resource, infrastructure could be leaked/orphaned. Now the create/destroy
options are provided with additional options which enable the destroy command to
tear down infrastructure even if the HostedCluster doesn't exist. This increases
the chance we'll cleanly destroy all infrastructure when test failures occur.