Dynamically discover release image OS metadata
Before this commit, CoreOS metadata was hard-coded in this Git repo and required
manual maintenance to map OCP release versions to compatible AMIs. Now, the
metadata is discovered automatically as part of the general release image
metadata lookup, as the CoreOS metadata is part of the payload as of
https://github.com/openshift/installer/pull/4760.

After this commit, HyperShift will only be compatible with OCP release images
which embed the CoreOS metadata.