allow KAS-provided or externally managed OIDC service account issuer discovery
This commit replaces the current s3-based OIDC endpoint with the one
provided natively by the kube-apiserver through the
ServiceAccountIssuerDiscovery feature.

It does this my exposing a new Route to the kube-apiserver on the
Hosted Control Plane with the OIDC service type Route is specified,
which is the default. The host for this Route will be the issuer URL
registered with the OIDC provider created in IAM.