Fix node pool controller panic and release image API validation
Before this commit, the `release.image` API field of hostedcluster and nodepool
was incorrectly allowing empty string values for the release image. Then, some
code in the nodepool controller would panic trying to default the missing value
due to a missing `nil` check waiting for hostedcluster status.

This commit cleans up the mess by updating the API to validate the `image` field
be non-empty. The CLI code is updated to set the required value on nodepool, and
the controller-level defaulting is removed entirely as the value is the
responsibility of the client.

The `create nodepool` command now accepts a `--release-image` argument and
defaults to use the same value as the referenced HostedCluster.