HCP: add ingress global config
The CLI create IAM now needs the Ingress global config to be set on a
cluster in order to determine what the OIDC route URL will be.

We currently create the default IngressController config, which allows
ingress in the guest cluster to work as expected, but not the global
config.

This PR sets the global config to enable nested Hypershift clusters.