hypershift-operator: Ensure owner refs on controller-generated secret
The hypershift-operator generates an admin kubeconfig secret in the
hostedcluster namespace for interfacing with the hosted control plane.

Before this commit, deleting the hostedcluster orphaned that secret, which
users don't manage. This commit introduces an owner reference from the secret
to the hostedcluster so that when the hostedcluster is deleted, so is the
secret. Otherwise this is basically a resource leak.