Refactor core control plane component reconciliation
What does this do?
It modifies the way that Kubernetes core components are reconciled by
the control plane operator by replacing a set of manifests with go code
to generate and reconcile resources.

Why is it needed?
- Using go templates to render control plane components makes it hard to
understand the relationships between the different resources needed for
the control plane. Changes in one manifest that may affect a different
manifest are not immediately obvious. The approach of using go code
makes relationships easier to navigate and understand.
- As we support multiple versions of the control plane and allow
configuration knobs in the control plane, we may need to actively
reconcile what is there with new changes coming in and block changes
that are not supported or invalid.  This approach allows us to integrate
that logic into reconcile functions.
- By splitting out reconciliation functions, we are better able to unit
test reconcilation logic.

How is the code organized?
This PR creates several packages under the `hostedcontrolplane` package
of the hosted control plane controller. There is a manifests package
that contains functions to locate resources reconciled by the hosted
control plane controller. Packages that corresponding to different
components (such as kas, kcm, scheduler) contain reconcile logic for
resources in those components.

What remains to be done?
Functions that reconcile resources are still using a parameters resource
as state without making inputs to reconciling a resource explicitly
clear. These functions should eventually be made stateless.
Several components still need to be converted over from
templates to go code (openshift apiserver, controller manager, etc).