cli: Support user-provided hostedcluster annotations
This commit adds an optional `--annotations` flag to the `create cluster`
command which allows the user to provide annotations which will be
added to the hostedcluster. The flag can be passed multiple times to
add many annotations.

This is useful for things like associating a CI-generated cluster with
its CI metadata (Prow URLs, etc.)