Drop ExternalInfraCluster CRD in favour of AWSCluster
This PR drops the ExternalInfraCluster CRD in favour of AWSCluster.

Originally we added support for externally managed infra in CAPA via the ExternalInfraCluster CRD https://github.com/kubernetes-sigs/cluster-api-provider-aws/pull/2124 and we used that commit of CAPA in hypershift.

Later on we decided to revert that approach upstream and reuse the existing ${infra}Cluster CRDs with an annotation to support externally managed infrastructure https://github.com/kubernetes-sigs/cluster-api/pull/4135

This PR bring latest CAPI/CAPA.

As a follow up we need to rebuild the images and consume them from quay.io/hypershift.