e2e: Refactor and improve diagnostics
This commit incorporates a new diagnostic step during teardown which dumps key
resources from the guest cluster's API server if available. This enables some
baseline level of post-mortem analysis of the cluster. In some cases it's
possible to actually use `oc adm must-gather` against the hosted cluster, but
only if it is known that pods can be reliablu scheduled, which is often not the
case in common failure scenarios.

The resources are dumped to a subdirectory of the artifacts directory along with
the usual dump obtained from the management cluster. Logging is also
consolidated throughout the test code to use the logr interface.

In addition, much of the e2e code has been refactored to try and make it easier
to share boilerplate test code.

One other notable change in this commit is that the various command structs are
updated to stop receiving AWS client interfaces: passing these through
everywhere has proven error prone in calling code, and the benefit of sharing a
client instance across logical command contexts is unclear; instead, the code
now passes around AWS credentials and allows the commands to build their own
clients internally. As an additional improvement here, each discrete AWS client
now gets its own unique agent identifier to help contextualize requests
(something which was not possible when sharing clients.)