Introduce OpenTelemetry tracing support
This commit adds experimental support for OpenTelemetry tracing. A new
`--otlp-endpoint` option is provided — if specified, traces will be exported to
this endpoint (which must be an OTLP gRPC endpoint). By default, traces will not
be exported anywhere.

This commit also adds some basic tracing to the nodepool controller as an
example to build upon.