cli: always create a NodePool with the desired replicas
Before this commit, `create cluster` didn't create a NodePool if the
desired node pool replicas value is zero. However, a NodePool with
zero replicas is a valid and sometimes useful state.

This commit changes `create cluster` to always create a NodePool even
if the value is zero. To skip creating a NodePool entirely, use a replica
value of -1 instead.