WIP: cpo: refactor status handling
Before this commit, status handling in the hostedcontrolplane-operator was
inconsistent and ad-hoc. This commit tries to improve status handling by:

- Adding more granular condition types for various reconciliation tasks which
are then composed into higher level conditions (e.g. computing the
hostedcontrolplane Available condition as a function of both KAS and etcd
availability).
- Decoupling some status calculation from reconciliation so that status can be
updated indpendently of the ability to make progress in the reconcile loop.
- Using upstream condition types and utilities for working with conditions
- Consolidating status updates into a single deferred call so updates can be
collected along the way and persisted once with a patch, letting the tooling
deal with diffs and transition times.