Expose API for upgrades in NodePools
This abstract the API for upgrading NodePools as follows:

- Upgrade type: Replace / InPlace.
  - Upgrade strategies:
	Replace:
      - RollingUpdate.
      - OnDelete.

	InPlace:
      - Not Implemented.

UpgradeType defines if an upgrade happens by replacing a Machine or by running software in a exsiting Machine.

Strategy defines an algorithm (control and pace) for an upgradeType to take place across all the Machines in a NodePool.

This model provide us with the ability to implement new strategies at any point in time for both types of upgrades InPlace or RePlace.