Fix version reporting in control plane upgrades and add e2e tests
Before this commit, the version of the hostedcluster's release image was
persisted exactly once for the life of the hostedcluster, causing the version to
be misreported following a rollout to a new release image.

The root cause of the problem is a bug in the control plane operator which
causes the hostedcontrolplane controller to only ever update the version on
status if it was never set before.

This commit fixes the problem by ensuring the version on hostedcontrolplane
status is updated atomically along with the corresponding release image driving
the rollout.

This commit also introduces expanded e2e test coverage for upgrades to perform
deeper inspections of the cluster operator and hosted cluster versions to assert
the versions eventually converge following an upgrade.

Note: The cluster operator version assertions in particular are probably not
appropriate to assert directly in the tests and instead should probably be
considered by the control plane operator and rolled up into availability/upgrade
status later on. Then, the tests could rely on condition checks to understand
when the operators are ready instead of probing the guest api server (which I
believe is a layering violation.).

In the meantime, these tests are better than nothing and caught the bug.