Create a separate oauth server serving cert
What this does: Adds a new PKI serving cert for the openshift oauth
server. By default, it duplicates the content of the ingress cert.

Why is it needed: For ROKS in which their own certs are generated, the
oauth serving cert should not be synced into the guest cluster like we
do with the ingress cert. This allows them to generate separate certs
for each. The reason we use the same cert by default as the ingress cert
is that the console validates its connection to the oauth server using
the default ingress cert. There is no setting on the console to use a
different cert.