Improve reliability of release version rollout tracking
Before this commit, a new release image rollout version was assumed to be
successful without any meaningful invariants being enforced. As a result,
observers were given inaccurate information about the current state of the
rollout (e.g. reporting rollout success when the underlying CVO rollout is
incomplete or failing).

Now, when the `hostedcluster` reports that a release version is `Completed`, the
user can have more confidence that the version has actually been rolled out all
the way through to the hosted cluster's OCP components.

To achieve this, the hosted control plane operator now inspects the underlying
CVO of a hosted cluster and incorporates the CVO's `clusterversion` state into
the `hostedcontrolplane` rollout tracking functions. For example, when rolling
out a new release image for a `hostedcontrolplane`, completion will only be
reported once the `clusterversion` inside the hosted cluster reports having
completed rollout of the semver for the release image.

As an additional piece of state to leverage for rollout analysis, the `Failing`
condition of the `clusterversion` is now mirrored to the `hostedcontrolplane`
status via the `ClusterVersionFailing` condition to provide a more authoritative
proxy for hosted cluster operator health in aggregate.

Internally, the hosted control plane operator is now capable of efficiently
inspecting hosted cluster resources as well as triggering `hostedcontrolplane`
reconciliation based on events originating from the hosted apiserver. This
capability is driven by a new intermediate caching abstraction paired with a new
cache management controller which knows how to reconcile hosted cluster
kubeconfigs and maintain an internal resource cache with event handlers. This
cache is shared with the hosted control plane controller and provides the hosted
apiserver client and event access.