WIP: Surface clusteroperator availability through hostedcontrolplane conditions
This commit adds a new `ClusterOperatorsAvailable` condition to hostedcontrolplane
which represents the aggregate status of the `Available` condition of clusteroperator
resources in the hosted control plane. This enables observers to assess the available
status from the hostedcontrolplane as a source of truth rather than requiring the
observer to reach in to the hosted control plane itself to assess health.

This is achieved by updating the control plane operator to actively query the hosted
apiserver to inspect clusteroperator status during each reconciliation.