Add pull secret to release info provider interface
Currently the releaseinfo provider interface only takes an image name to
lookup release info. However, we should be using the pull secret that
was provided for that release to pull that image. This commit adds a
pull secret parameter to the provider lookup function and updates the
implementation of pod provider to use the pull secret.