Extract API and control-plane-operator into Go submodules
This commit takes the first steps towards preparing control-plane-operator for
extraction into a separate Git repository. The first steps to achieve this are:

1. Extract the `api` package into its own Go module so that external consumers
don't inherit the operator's dependency graph
2. Extract the `control-plane-operator` into its own Go module so that it's not
entangled with any other module in this repository

This commit also extract a `support` module which has release-independent shared
library functions used by both the hypershift-operator and the
control-plane-operator.

The control-plane-operator also gets its own `Makefile` which is called out from
the root `Makefile` for now to maintain compatibility with existing workflows
during the transition phase.

Third-party APIs are also cleaned up and moved to be near the modules which use
them.

Note that changes to the `api` package will now require updating the root and
control-plane-operator vendor directories.