Revert submodules
Modules were introduced for https://github.com/openshift/hypershift/issues/426, but the
underlying assumptions have changed and we no longer intend to split out the control
plane operator repo, instead opting for an in-tree branching solution which totally
negates the need for separate modules for now. This commit reverts all the submodules
since in this branching paradigm they aren't required and only add complexity for now.