Consolidate e2e scenarios into a single test for now
The e2e tests run serially are too slow, but we can't yet parallelize them
because the log output becomes interleaved and unintelligable. Until we can
run the tests in parallel, consolidate the existing scenarios into one to
preserve the same coverage but with shorter total test duration.

This also temporarily disables e2e dumps which are hanging on `oc adm inspect`
for some reason. It's more important to fix speeds now, and we can diagnose
and fix the dump in a followup.