Minimum viable HA etcd API and implementation
This commit replaces the CoreOS etcd-operator with a minimum viable HA etcd
implementation based on StatefulSet. In this implementation, etcd is deployed
directly by the control plane operator into a StatefulSet with a static
configuration that provides resiliency against connectivity outages and process
restarts, but does not protect against persistent data loss of any member.
Recovering from data loss will be addressed in subsequent work.

The following is a summary of the assumptions and design of the approach.

Given the following invariants:

1. etcd member addresses are predetermined and immutable (via StatefulSet
service)

2. etcd cluster size is immutable

The following holds true:

1. Using a static configuration, and assuming no persistent volume data loss,
the etcd cluster can be bootstrapped and survive arbitrary member process
restarts and network outages

2. etcd cluster membership reconfigurations are only required under the
following circumstances:

    a. Minority of members suffer data loss resulting in new member identity
      - Remediation: delete members' old IDs from etcd cluster to enable
        rejoin/seed

    b. Majority of members suffer data loss resulting in new member identity
      - Remediation (lossy): restore member data directories from snapshot

Altogether, these are the implications:

1. Core operational logic can be reduced to deploying etcd directly into a
StatefulSet

2. Data loss detection and remediation can be considered a related but
orthogonal concern; for example:

    a. A future control plane operator controller could detect minority members
    who have lost their data/identity and handle removing the underlying members
    from the etcd cluster to enable join/seed

    b. A future control plane operator controller could detect majority member
    data loss and deal with remediation based on a HostedCluster policy
    configuration, e.g. "manual" (inform the user and let them handle a restore
    operation), "automated" (restore snapshots in the defunct pods)