Fix mco/mcs script bug that renders master ignition
The script to run mco in bootstrap mode uses 'exec' to call the mco
binary. The exec causes the mco process to take over the main container
process and when it exits, the container dies without executing the rest
of the script which swaps master/worker configs.

This PR fixes that problem and properly labels core machine configs with
the worker pool instead of the master pool.