Controlplaneoperator: Don't set status.kubeconfig before it exists
Currently, creating a cluster always results in a bunch of errors from
the hosted cluster reconciler because it got a 404 trying to fetch the
hosted controlplane kubeconfig. This happens because the hosted
controlplane operator sets status.kubeconfig before that secret actually
exists, whereas the hosted controlplane operator assumes that as soon as
that field is non-nil, it exists.