Ignore Duplicate permission error code when creating security group
Create infra is sometimes terminating with an error because it claims
that the security group rule we are trying to create is a duplicate.
This is likely caused by retry logic in the AWS client. This change
avoids the error by ignoring an error of duplicate permission when
creating security group rules.