Delete NodePool secrets fetching by annotation
Before this commit if a NodePool was deleted before having the nodePoolAnnotationCurrentConfigVersion annotation then the token and userdata Secrets would be leaked as the value of this annotation was used as part of a naming convention to create and find the secrets.
This PR let the NodePool controller fetch the secrets that are annotated with the nodePoolAnnotation instead which eliminates any possible race.